Several improvements
Added a function that allows you to activate and deactivate dark mode in the UI. A setting has also been added to the UI that allows you to preselect the event filter. A bug has also been fixed that caused the icon to not load correctly for notifications (if the server is not directly accessible from the client).